Added resonance presets
I added in the choice for the user to use their entered values, or some preset values. One is resonance (oscillations grow due to forcing function) and one is anti-resonance (for lack of a better term, oscillations shrink due to forcing function, not the damping coefficient).
Also the forcing function is now a dropdown, it doesnt have anything yet but should later be populated with some interesting forcing fuctions. This will then be translated to a functionNumber, which can be used in the numerical calculation to determine which forcing function to use.