A few additions
First off, changed the forcing function value getter in spring.py as we didn't need the parameter for fNum since we store self.funcNum anyways.
Also, the simulation now displays the corresponding ODE.
Finally, made it so that if the mass reaches equilibrium before the 10 seconds is up, it stops the simulation.