Read the description - error checking
So basically, in spring.py, where it says yAct[i] = ... in the analytical() function, enter the analytical solution to the ODE that you're trying to find the error on. Then, launch the simulation. BE SURE THAT YOUR PARAMETERS MATCH THOSE USED FOR THE ANALYTICAL SOLUTION. This will plot the numerical and analytical solution on the same plot, and then it's going to print to the file errorData.txt the relative error between solutions at each point. NOTE: the first point is at time t=0.01, the second is 0.02, and so on, up until 9.99 (if you did 10s simulation).